move waypoints graph component to a private system component
